[Others] 관리자 토너먼트 유저 추가 검색창 수정 #1175 (#1201)
* [Refactor] useQuery staletime 추가 #1175

* [Refactor] 콘솔 로그 제거 #1175

* [Fix] 아이디 유효성 검사 #1175

* [Refactor] AdminSearchBar로 대체 #1175

* [Style] 토너먼트 유저 수정 모달 디자인 수정 #1175

* [Chore] 쓰이지 않는 훅 삭제 #1175
