+++ styles/admin/modal/AdminTournamentParticipantEditModal.module.scss
@@ -10,6 +10,7 @@ $notJoinedColor: grey;
   height: 30rem;
   overflow-y: auto;
   color: black;
+  background-color: rgb(255, 246, 234);
 
   .deleteInputTitle {
     font-size: 0.7rem;
@@ -28,16 +29,8 @@ $notJoinedColor: grey;
     justify-content: space-between;
   }
 
-  .stickyHeader {
-    position: sticky;
-    top: 0;
-    z-index: 1;
-    display: flex;
-    padding: 0.5rem;
-    background-color: white;
-    border: 1px solid grey;
-    border-radius: 1rem;
-    justify-content: space-evenly;
+  .searchBarWrapper {
+    padding: 0 2rem 1rem;
   }
 
   h2 {
@@ -49,8 +42,14 @@ $notJoinedColor: grey;
   .hr {
     width: 100%;
     height: 0.5px;
-    margin-bottom: 1rem;
     background: black;
+  }
+
+  .tournamentUserList {
+    height: 60%;
+    overflow-y: auto;
+    border: 1px solid grey;
+    border-radius: 1rem;
   }
 
   li {
@@ -82,7 +81,7 @@ $notJoinedColor: grey;
     }
   }
 
-  .buttonGroup {
+  .cancleButton {
     position: sticky;
     bottom: 0;
     display: flex;
